Change Damage Event to match new server schema

+++ Wow-Raid/Wow-Raid/LogClasses/DamageEvent.cs
@@ -14,7 +14,7 @@ namespace Wow_Raid.LogClasses
         int encounter;
         int logno;
         int damage;
-        double timestamp;
+        long timestamp;
 
 
         String source;
@@ -26,7 +26,7 @@ namespace Wow_Raid.LogClasses
             this.encounter = (int)row["encounter"];
             this.logno = (int)row["logno"];
             this.damage = (int)row["damage"];
-            this.timestamp = (double)row["timestamp"];
+            this.timestamp = (long)row["timestamp"];
 
 
             this.source = (String)row["source"];
